Inserir e alterar produto vinculo com item de composição funcionando

+++ Código/Restoran/src/Bean/ItemComposicaoProdutoBean.java
@@ -110,7 +110,7 @@ public class ItemComposicaoProdutoBean {
 	}
 
 	public String chamadaAlterar(Integer id) {
-		itemComposicaoProdutoAlterar = fachada.ItemComposicaoProdutoBuscarPorId(id);
+		itemComposicaoProdutoAlterar = fachada.ItemComposicaoProdutoBuscarPorId((long) id);
         return "alterar";
     }
